conexion to database succefull and change type of id to bigNumber to User model

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/com/peryloth/r2dbc/MyReactiveRepository.java
@@ -4,7 +4,9 @@ import com.peryloth.r2dbc.entity.UsuarioEntity;
 import org.springframework.data.repository.query.ReactiveQueryByExampleExecutor;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
+import java.math.BigInteger;
+
 // TODO: This file is just an example, you should delete or modify it
-public interface MyReactiveRepository extends ReactiveCrudRepository<UsuarioEntity, Long>, ReactiveQueryByExampleExecutor<UsuarioEntity> {
+public interface MyReactiveRepository extends ReactiveCrudRepository<UsuarioEntity, BigInteger>, ReactiveQueryByExampleExecutor<UsuarioEntity> {
 
 }
